requirements.txt
torch>=2.1.0
pettingzoo>=1.24.0
gymnasium>=0.29.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0
